add tests to TraverseEngine with RedisMongo
