+++ tests/integration/test_client.py
@@ -1,13 +1,17 @@
 import pytest
 from hyperon_das_atomdb.utils.expression_hasher import ExpressionHasher
-from .remote_das_info import remote_das_host, remote_das_port
+
 from hyperon_das.client import FunctionsClient
+
+from .remote_das_info import remote_das_host, remote_das_port
 
 
 class TestVultrClientIntegration:
     @pytest.fixture()
     def server(self):
-        return FunctionsClient(url=f'http://{remote_das_host}:{remote_das_port}/function/query-engine')
+        return FunctionsClient(
+            url=f'http://{remote_das_host}:{remote_das_port}/function/query-engine'
+        )
 
     @pytest.fixture()
     def node_human(self):
@@ -67,10 +71,10 @@ class TestVultrClientIntegration:
         assert result['targets'] == [node_similarity, node_human, node_monkey]
 
     def test_get_node(
-        self, 
-        server: FunctionsClient, 
-        node_human: str, 
-        node_monkey: str, 
+        self,
+        server: FunctionsClient,
+        node_human: str,
+        node_monkey: str,
     ):
         result = server.get_node(node_type='Symbol', node_name='"human"')
         assert result['handle'] == node_human
@@ -90,15 +94,19 @@ class TestVultrClientIntegration:
         node_mammal: str,
         link_similarity_concept_concept: str,
         link_inheritance_concept_concept: str,
-        node_similarity: str, 
+        node_similarity: str,
         node_inheritance: str,
     ):
-        result = server.get_link(link_type='Expression', link_targets=[node_similarity, node_human, node_monkey])
+        result = server.get_link(
+            link_type='Expression', link_targets=[node_similarity, node_human, node_monkey]
+        )
         assert result['handle'] == link_similarity_concept_concept
         assert result['named_type'] == 'Expression'
         assert result['targets'] == [node_similarity, node_human, node_monkey]
 
-        result = server.get_link(link_type='Expression', link_targets=[node_inheritance, node_human, node_mammal])
+        result = server.get_link(
+            link_type='Expression', link_targets=[node_inheritance, node_human, node_mammal]
+        )
         assert result['handle'] == link_inheritance_concept_concept
         assert result['named_type'] == 'Expression'
         assert result['targets'] == [node_inheritance, node_human, node_mammal]
